requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
passlib==1.7.4
py==1.11.0
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
